db: Print query time with --debug-sql.
With --debug-sql, it was difficult to determine just by monitoring output if a particular query was slow, or if tup is busy doing something else after printing the query line. Now the transaction_check saves off the query string in a temporary buffer, and the buffer and actual query time is displayed when sqlite3_reset() is called (through the msqlite3_reset() wrapper).