parser: Print errors to tf->f instead of stderr where possible.
Sometimes error messages would get dumped outside of the actual parsed Tupfile heading, making it difficult to see what the problem is. This puts most of the error messages inside the temporary output file for the particular Tupfile so they can all be displayed under the node heading. There are a few that still output directly to stderr - those associated with creating/closing the temporary output file itself, and a few memory allocation ones in helper functions that shouldn't ever fail anyway.