parser: More accurate time reporting.
When tup sees an input from another directory, it stops parsing the current Tupfile and parses the dependent Tupfile. However, the time to parse the second Tupfile would essentially be included in the first one as well because the timespan comprised the whole range of time to parse both. This is potentially confusing, since a Tupfile that is dependent on many other directories could report a large parsing time, even if the actual time is quick. This change subtracts out the time to parse any dependent Tupfiles by adjusting the start time.