Added t4069, t4070 - support for ^c flag for chroot.
This allows a Tupfile to specify that a certain command should be executed in a chroot environment, without having to run everything with full_deps enabled. Mostly this is to support gcc with the --coverage option - without it, gcc will write the @tupjob path into the executable, which doesn't exist when the program runs.

The downside is the somewhat awkward way that it gets specified, but the only other way to get the information to the updater is if there is an extra column in the database, and it still needs to be set somehow in the actual :-rule anyway. Since we already use the ^ TEXT^ area as a weird way of passing metadata to the updater, and there was already space for ^-flags, it seemed like a reasonable fit.