windows: Don't inject into mspdbsrv.exe
The mspdbsrv.exe is spawned by cl.exe if you generate the pdb output, but the process runs as a detached server process in the background after cl.exe returns. If we inject into it, then it will hang on to our deps & output files so we can't remove them.