monitor: Always watch file events, even during an update.
We should be able to catch events if a user creates a file while an update (including autoupdate) is taking place. This especially helps with autoupdate while doing a lot of file accesses (such as untarring a large zip) to ensure that tup does not skip any files. The monitor just ignores events on generated files if the timestamps match, so that it doesn't complain that "foo.o" was modified, since it was the updater doing the modification.