Added t7053, t7054 - use stored environment in autoupdate.
When the monitor process starts up, it has whatever environment the user had when first running it. If they wish to change the environment later, they can do so by modifying their environ and running 'tup upd'. However, with autoparse/autoupdate, the monitor would keep trying to revert to the old environment. Instead, it is better to have the monitor just use the cached values for the environment, so that the user can modify them without having to restart the monitor.