Fix t7052 race conditions.
It is possible that the monitor starts scanning after a file is created, which would trigger an initial autoupdate. This could then be followed by a second autoupdate, resulting in the test failing. Additionally, sometimes the .monitor.output file isn't flushed by the time we check, so if we fail the initial test just sleep a bit and try again.