Database version 14: Add config_list for tup.config changes.
Previously, there was a unique directory node, VAR_DT, which had a tupid of 2. Various parts of the code would know to check for or ignore this particular value, since its presence in the create_list meant that we had to reparse the tup.config file for new @-variable changes. With variants, we would like to potentially have multiple tup.config files, each with their own @-variable namespace. As such, we only want to parse a particular tup.config file if it has changed, which we can't discern from a single tupid value of 2. Instead, each tup.config can be treated as a virtual directory with @-variables as children, and the tup.config file's tupid can be put into the new config_list that comprises the first phase of processing. After this, all create_list nodes can be parsed, and then all modify_list nodes can be executed.

Additionally, the node table is modified to have a srcid value. This will be used to point from variant directories to the corresponding directory in the src tree.