Added t4072 - test for /proc/self access in a chroot.
The /proc/self symlink points to the current process's pid. However, when we access it in a chroot, the actual readlink() call goes to our fuse process, so the kernel would assign it our pid rather than the sub-processes pid. We can bypass the kernel here and just write in the sub-process's pid directly.