Added t8014, t8015 - can't use files cross variants.
Since variants are optional, it doesn't make sense to try to use a file from one variant in another. The check for this is done in both the parser stage (so errors can be given more quickly) and when input files are checked (in case the file isn't mentioned in the Tupfile).