fuse: Print filename if unlink() is rejected.
When tup denies an unlink() call because a sub-process tried to unlink a file that it did not create, the error message should contain the actual filename. It is possible that the sub-process won't display an error message itself, and without the filename it is hard to track down the origin.