Added t8016 - go from in-tree to a variant.
I was trying to rely on the fact that calling the parser with a disabled variant would clear out the in-tree build when switching from in-tree to variants, but unfortunately the in-tree build files are not removed until after all Tupfiles are parsed. This means a variant Tupfile could be parsed and see generated files from the in-tree build.

The fix here is to detect when the in-tree build is going away, and wipe it out before doing any parsing. This unfortunately requires going through the whole database, but is necessary to ensure that the in-tree build does not affect the variant.