Added t8018, t8019, t8020 - tests for cleaning variant dirs.
The variant build directories should be cleaned up automatically as Tupfiles are removed in the src directory. This is a little tricky since we have to wait until delete_files() (variants won't be loaded during the scan / monitor stage). Then we check if the directory can be cleaned up with two quick queries.