Added t8021 - variant -> in-tree without a tup.config
Even if we don't have a top-level tup.config, if we remove all the variants, it should fallback to an in-tree build. This is done by reparsing all Tupfiles for the in-tree variant. Note that we cheat and know whether or not to re-parse everything by whether or not our ghost in-tree tup.config exists. If it doesn't, it means it's our first update and so we don't have to issue the O(n) db query.