Make sure we don't reparse all when we don't need to.
We need to know when we go from a variant-based build to an in-tree build. However, this is not possible to do just based on variant_list because that can be cleared out by the scanner/monitor. Instead they need to note that variants have been removed, so that if we are creating an in-tree variant we can re-parse everything. Of course we don't want to issue the re-parse all command everytime the variant list is empty, because we could just be starting up for the first time.