Variants currently disabled in Windows.
The file-tree changes that allow Windows to use .tup/tmp for all output files needs to be merged before variants will work. Currently Windows writes files directly to the srctree, so they can't easily be moved to the variant tree like on Linux/OSX. This will be reverted after that branch is merged.