Added t8004, t8005 - tests for variants.
Variants currently function by creating a top-level subdirectory with a tup.config in it. The src directory tree is then duplicated here, and all new sub-directories get their own tup_entrys and are parse-able. The "default" variant, which is the standard in-tree one, is disabled if other variants are used. Otherwise when a variant overlays the src file-system into its variant tree, it will see the generated files from the in-tree variant and get confused.