Added t8006 - make sure we can parse a variant with a ghost.
If a variant tries to read from 'foo.h', then we can get a ghost in build/foo.h even if we end up reading foo.h from the src tree. In this case, we need to still be able to parse the Tupfile when our input is found as a ghost in the variant.