Added t8029, t8030, t8031 - check for generated files overwriting src nodes.
Normally when we create a generated node in the parser, the database makes a check to see if we are converting a ghost node into a generated file, or if we are attempting to stomp an existing regular file. However with variants, the generated file will be going in a different directory, so we don't get the useful error mesasge about overwriting a regular file. Without this check, we will just fail in the update step as a result of trying to access the regular file in the source tree.