Added t8027 - correctly if tup.config is a symlink.
Tup was expecting tup.config to be a normal file, but it may be useful to let it also be a symlink. This way the user could have a directory of configs for their project to support variants, but not force an end-user to build with those variants by checking in all of the build directories. The downside is now when the file is opened, tup has to start a server just to read the file. In order to keep track of it, tup adds a link directly from the read files to the tup.config node in the database. An extra query is used in tup_file_mod to add tup.config to the config_list as appropriate.