Added t8033 - handle the user removing part of a variant.
If a user does an 'rm -rf build/sub', then we should re-create the subdirectory and re-run all those commands. This is handled by detecting that a variant directory was removed and adding the corresponding srctree tupid to the create_list. This will cause tup to re-generate the directory and re-parse it, running all the commands for that directory again in the process.