Added t8034 - rm -rf a variant directory with subdirectories.
When we detect that a variant directory has been removed, we add the srctree directory to the create_list if we aren't doing a forced delete (ie: the user has removed the directory). However, the tup_db_delete_dir function assumed we would always do a forced delete, so the corresponding subdirectories wouldn't necessarily also be added to the create_list. This change passes along the force information in tup_db_delete_dir so that all removed variant directories will be handled properly.