Don't duplicate node creation logic in get_rel_tent().
The get_rel_tent() function was re-doing some of the work already handled in tup_db_create_node_part(). Specifically it was searching if a node existed, and if not it would try to create it. Then tup_db_create_node_part would also see if the node existed, and if not try to create it. Both functions were also checking for ghost nodes, which is unnecessary and confuses the responsibility of who should be doing the checking. Now get_rel_tent() just lets the db do the work, and a flag is used to determine whether or not we need to create the variant directory.