Added t8035 - remove part of a variant and the srctree at the same time.
When we detect that a variant directory is removed, we add the corresponding srctree into the create_list so the directory can be re-created if necessary. However, it is possible that the srcid is no longer valid (either the node no longer exists, or it is a ghost). In this case, we don't actually want to add the id to the create_list.