Clear the variants_removed flag.
There doesn't seem to be a way to cause the reparse_all to trigger incorrectly even if the variants_removed flag is left set. This is because the once we are in-tree, the root tup.config node is in the variant list so we never set the new_in_tree flag. However, it seems good practice to clear it out after it is used.