Added t8036 - rm part of a variant with the monitor running.
The monitor needs to load the variants itself, which it does during its initial scan and after any update completes. This allows it to correctly set the create flags when a variant subdirectory is removed. For the updater, it still needs to load variants even if it is not scanning so that it can process those create nodes correctly.