Added t8037 - move a src directory with a variant build.
When we move the src directory, the Tupfile that tries to include files from that directory should fail. However, with the post-parser cleanup of variant directories, this wasn't happening. Instead we need to clean up old variant directories before parsing anything. We can do this by checking if the srcid is a ghost, and if so, remove the variant directory.