Remove variant dir reclaimation code.
Variant dirs were being cleaned up after the parser stage when the generated files were deleted. However, this caused problems if the variant dir was still used by another Tupfile (t8037) so instead variant directories needed to be cleaned up before-hand. In order to clean up directories properly in t8020, we need to flag anything pointed to by a ghost Tupfile when a directory is removed.