Added t7055 - move a directory that contains a file included by a Tupfile.
The tup_db_modify_dir() function was recursively adding all files in a directory that was moved to the modify_list. However, there is extra logic in tup_file_mod_mtime() that knows to also add dependent directories to the create_list. We need to somewhat duplicate that info here - going through tup_file_mod_mtime() would be difficult because of the checks for deleting generated nodes, which we aren't actually doing (just moving them).