Added t8038 - move a srcdir with the monitor running.
This is similar to t8037, but we have the monitor running so the directory node gets renamed, rather than having the old directory tree deleted and a new one created. This means our variant dir still has a valid srcid that points to an existing node, but we know that our variant dir is no longer valid because the name of the node doesn't match.