Added t8039 - move multiple levels of srctree with the monitor running.
This is quite tricky to handle, since we get a single move event, and the monitor has to update all of its dircache entries after the updater changes all of the variant directories. A temporary list is used to get a list of the directories that will be cleaned up since a single variant subdir can wipe out multiple entries in the create list as it recurses down.