Added t8044 - clean up directories as variants are created.
If we create a new directory for a variant but don't put a tup.config in it, then that directory is cloned in all active variants. When we later add a tup.config, those cloned directories should be removed so that we have the same structure in the case where we add a variant directory with a tup.config in it in a single update.