Added t8047 - accidentally put tup.config in the wrong directory.
This is a bit tricky since we clean up variants when tup.config is removed. However, that should only be done if it was ever actually considered a variant in the first place. This moves the variant deletion logic from when tup.config is removed to process_config_nodes(). This is done by just making the tup.config node a ghost (like in the root variant case). This also removes the need for the variants_removed config variable.