Added t8048, t8049, t8050, t8051, t8052, t8053.
These tests check the 'tup variant' helper function that can be used to create a bunch of variants from a directory of config files. The idea here is that a team can check-in config files for their supported configurations, then when a user does an initial check-out they can do 'tup variant configs/*.config' and have all the variants created for them, rather than creating each one manually. The actual tup.config file in the variant is symlinked to the configs directory, so any changes will be reflected in the file that is checked in to version control.