Added t8054 - try to rm -rf multiple variants.
When the scanner goes to clean up files in tup_db_scan_end, it is essentially going through the missing nodes in no particular order and calling tup_file_missing(). If it called tup_file_missing() on the tup.config node first, it would work, but if it called tup_file_missing() on the build directory first, the recursion would wipe out tup.config completely instead of making it into a ghost for cleanup. Creating many variants and removing them triggers this behavior.