Added t8055 - make sure variant srcid is not -1.
If we have a variant, then remove it and just do a scan before re-creating it, the srcid gets reset to -1 when the node is converted from a ghost, but we weren't setting the srcid again since it wasn't a "new" variant. This makes sure the srcid is always set for the variant directory.