Reorganize scanner logic.
During a scan, every file-based node (TUP_NODE_FILE, TUP_NODE_DIR, TUP_NODE_GENERATED) must be loaded in the DAG. Tup was also putting each of these tup entries into a tree, so that after the scan we could see which files were missing. However, this caused some issues since the order that the missing nodes are processed is not specified. For example, if a directory was removed, either the file in the directory or the directory itself could be processed first. This changeset instead creates a smaller tree of nodes for each directory, and this tree is processed after the directory is scanned. This means if a whole sub-tree is missing, only the top-level node will get a tup_file_missing() call. This makes it much more similar to the monitor case of moving a directory out of the tup hierarchy.