Update t8054 - don't reparse everything when rm -rf'ing a variant.
When we remove part of a variant, we need to reparse the src directory so that directories can be recreated as needed. However, if we remove the entire variant, we don't want to trigger this logic. Now there is a check in tup_file_missing to see if we are removing a variant build directory so that the force flag can be set appropriately. This relies on the fact that we now only call tup_file_missing() on the first directory of a missing sub-tree.