Added t8061 - use include_rules in a variant build.
The include_rules directive works by going all the way up the tree with "../" and appending "Tuprules.tup" to the string. In a variant, we only want to go up to the top-level build directory, not all the way down to DOT_DT. Otherwise we will try to include the top-level Tuprules.tup file twice.