Update t8033 - get a warning if a variant directory is removed.
Normally we warn if generated files are removed. We don't get warnings if a whole sub-tree is removed though since the directory's entries are removed with force=1. These generated files couldn't be re-generated anyway if the directory with the Tupfile is gone, so that is the desired behavior. With variants, however, we could remove generated files by removing just a variant subtree. In this case we still need to delete with force=1 so the nodes are removed, but we may be re-generating them anyway if the src tree hasn't been deleted. We can still produce a warning about the variant directory being removed, so that is what t8033 has been updated to check.