Added t8062, t8063 - move the variant build directory to the src tree.
This has slightly different behavior depending on whether the scanner or the monitor is running. For the scanner case, we see it as a removed variant and new source files, so all of the generated nodes become regular files. For the monitor, since we use tup_db_change_node(), the variant_list now points to a variant that is not at the root directory. Tup now detects this as a removed variant and wipes out the build tree, similar to if tup.config had been removed. In either case, this is silly behavior on the part of the developer - the important thing is that tup doesn't break if it happens.