monitor: Run valgrind/helgrind on the monitor process.
In the test cases, the 'update()' function will run valgrind or helgrind if requested. However, the monitor was going unchecked since we called 'tup monitor' directly. Now there is a corresponding 'monitor()' function that will run valgrind/helgrind on the monitor process. The eotup() function calls stop_monitor() if necessary.