monitor: Don't re-open the top of tup in stop_monitor().
When starting up the monitor, it calls stop_monitor() to kill any previous monitor process. We've already called tup_init() at this point, so we don't need to re-search up the tup hierarchy and open the top-level file descriptor. This code is only needed for the explicit 'tup stop' command, which doesn't go through tup_init().