monitor: Free events only after going through the whole list.
When a file is moved, the monitor stores an IN_MOVED_FROM event and an IN_MOVED_TO event. The 'to' event then points to the 'from' event. We need both around when processing the 'to' event, so if we free the 'from' event too early then the pointer is no longer valid. This is solved by just iterating through the list to handle each event, and then freeing them all later.