Check return value of the monitor in the test cases.
Now the test scripts run the monitor without daemonizing, so we can get the pid and wait on it. This lets us collect the valgrind results and return a failure if something is wrong. There is now a new 'waitmon' command to wait for the monitor to start up, since we won't be blocking the shell script until the monitor is daemonized as is the case when we run it in the background. Also, the monitor is now shutdown with SIGHUP instead of SIGKILL so that we can exit the monitor loop by setting a signal variable.