Fix t7037 race conditions.
In t7037, we create a directory and then promptly remove it. A check was added to see if the fstatat() in watch_path returned ENOENT, and if so to skip the file. However, for directories we then also go with an inotify_add_watch(), and an openat(), which may also fail if the fstatat() succeeds and then the directory is removed. This can be triggered by adding sleeps before the inotify_add_watch() or the openat() call, as well as in the test itself. It is a bit difficult to write a test case specifically for this behavior, but can happen occasionally.