Added t6065 - failed nodes shouldn't break things.
If a command fails, tup doesn't call pop_node() so that any dependent nodes are not executed (eg: we won't call the linker if one of the compilations it needs has failed). However, tup was removing the node from the graph, though the links were not removed. When destroy_graph() was ultimately called, tup would then remove the links from the other end where the node still exists, and try to access the freed memory. This resulted in a strange "fuse internal error", presumably because fuse ended up with an old pointer of tup's.

The fix is to just leave a failed node on the node_list, so that it will get freed in destroy_graph() along with all the rest. Since it was already cleared of all incoming links, it won't go back to plist so we don't have to worry about running it again.