Different way of implementing node vars: using relative paths from the root.
This avoids needing a bunch of logic to make node vars interact with variants correctly.
The paths generated when dereferencing a node var are slightly different, but still point to the same files.
Users of node vars shouldn't notice the change, unless they are saving the dereferenced path (which they shouldn't be doing).