win32: Make sure backup db is opened in binary mode.
In Windows, we wrap open() with our own implementation that goes through CreateFile, which doesn't do the newline conversion that open() normally does. However, we don't wrap creat(), so this was causing the backup database to be malformed. If we just use open instead of creat, then we go through the CreateFile route and can backup the database correctly.