Added t3043 - test for <groups>.
Now a command can specify a <group> in its output list, and its outputs will be associated with that group. Other commands can specify <group> in their input list, and any of the generated files associated with the group are acceptable for inputs. Essentially the <group> is used for the sticky ordering, rather than the files themselves. The files still get the actual normal links rather than the groups, so a group is not "modified" and does not automatically update all of the commands that use a group as an input.