Added t3046, t3047 - removing groups from commands.
We have to keep track of the old outputs and remove them from the group when the group is removed from a command. Also the group is now added to the ghost list for possible pruning on a commit when it is removed as an output. Pruning a group is slightly different than a ghost - here we just check whether there are any input or output links to the group. If there are, the group stays.