Added t3049 - make sure we can remove part of a group.
This test removes commands from a group one at a time to make sure the group sticks around until all of the commands are gone. A new check had to be added to tup_db_delete_links() to make sure that any groups pointed to by the command are removed, since that won't go through the existing oldgroup check. This is potentially somewhat inefficient, since we are selecting over the links twice in add_ghost_links() and add_group_links() before actually removing the links. Perhaps there are better ways of doing that.