Added t3050, t3051 - tests for group ordering.
A <group> in a graph should actually enforce ordering among commands. Currently this is done by expanding a group when we find it in the PDAG. This is potentially ineffecient since a group may link to many commands that we don't actually need to load. These extra commands aren't expanded and executed though, so the output is at least correct.