Try to use groups in a !-macro.
Specifying these is a little odd due to the way the !-macros are parsed. The exact text is saved until the macro is used, so we have to set a PROJ_ROOT $-variable that points to TUP_CWD and use this to get all of the !-macros to refer to the same group.