Added t6066 - don't break when converting a directory to a file.
When picked up by the scanner, the tup_file_mod_mtime() function would see that the file 'foo' is no longer a directory and call tup_del_id_type() on it to remove it from the database before inserting it as a new file. However, the ghost Tupfile in the directory would mean that the directory ends up as a ghost as well, and so trying to force the insert would fail. Instead we should check if the tup_entry still exists and convert it from a ghost if necessary.