Added t8071 - convert a directory to a symlink with a variant.
When the directory was deleted and replaced by a symlink, tup was not correctly removing the corresponding variant build directory. As such, trying to read from a symlink path was not actually reading from the symlink, but rather from the leftover variant directory. This would cause the dependency on the symlink to be missed.