fuse: Ignore first ENODEV error on FreeBSD.
There is a race condition between starting the fuse thread and making the first request in the fs. For some reason, the first one will fail with ENODEV in FreeBSD, but work after that. I also tried hooking the .init() callback in fuse and waiting until that was called, but that still had a possible race. The init() hook was a much lower probability failure, but easily seen by adding a sleep(1) in fuse_lowlevel.c after the call to f->op.init().

The fix here just makes an extra open on .tup/mnt, and ignores the error if it's ENODEV. We apparently only need to do this once - after the first failure, things work as expected. This is not just adding a delay, it seems to be causing fuse to go through its full initialization before returning the failure, so things should work fine after it returns.