fuse: Correctly peel the "/" path.
If a sub-process tries to access the top-level directory in fuse, the peel() function was returning "." instead of "/". This was a hold-over from the earlier path format where part of the directory came before the @tupjob marker. Now @tupjob comes before the full path, so if we get a stat for "/@tupjob-1", that means we should be statting "/" and return the result. Otherwise we are statting the current directory, which causes getcwd() in FreeBSD to break since that scans up the tree looking for a match of "/"'s inode and device descriptor.