FreeBSD: Add -ouse_ino to fuse.
This allows tup to set the st_ino field in getattr(), so that FreeBSD's getcwd() will work correctly. FreeBSD scans up the tree to look for a matching dev/ino to "/", so when we get to the top of the /@tupjob tree, we need to use the same dev/ino.