First work on adding lua to tup.
I managed to get tup to fully compile with lua on one computer.  None of
the tests have been run yet (or converted to lua).  All the original
parsing code is still in the source, with only the actual parse function
call being replaced with a the lua one.  Also, there's a lot of
unnecessary parsing in eval and such for variables, bins, etc. that is
not necessary (or useful) with lua Tupfiles.