win32: Updates to support newer mingw toolchain (with gcc 4.6)
For some reason many of the header files and macros have moved around or disappeared. One nasty change in particular has caused the stat() wrapper to fail. Since that is only used to find the fstatat()s for Tuprules.tup files, the open_notify() call is now in the fstatat wrapper.

Currently this allows the Windows version to build with both versions of mingw, but at somepoint the 4.5x version may be deprecated.