Fix color flags when using auto-detection.
In 390eb000 I broke the part that sets the enabled flags for stdout/stderr, so stderr was always disabled. This caused the command to no longer be highlighted in read when it fails. Unfortunately this is difficult to test automatically, since redirecting the output to a file causes a different command path to execute.