Added t8072, t8073, t8074 - generated -> normal file with variants.
This is basically t6053, but now we have a variant in the mix. With a variant we don't trigger the usual logic, which is just checking that the generated file has a newer timestamp when the rule is removed. Instead we have to look for a new file in the srctree to see if we should display an error message or not.