Added t8075 - manually created directories inside variants.
If a user manually creates a temporary directory inside a variant with a file in it, this shouldn't cause tup to complain. With a manual directory, the srcid is -1 since there is no corresponding source-tree directory, which tup was trying to check to see if the directory could be removed.