Added t2127 - use %d as a current directory name.
There was no easy way to get the name of the directory in the Tupfile, so if you wanted to name a library based on the directory name you would have to duplicate the information in the Tupfile. Tup already has this information though, so it is easy to expose using a %-flag. Note that only the lowest level directory is reported - it is not a partial path from the tup root, and certainly not a full path. It can just be used as a handy way to name things based on the directory name in macros.