win32: Fix t5077 to not track dependencies on directories.
The fuse fs was dropping dependencies on directories, but in some cases (FindFirstFile), Windows was not doing the same. This was causing problems because some commands were getting dependencies on common directories, so any changes to Tupfile or Tuprules.tup in those directories was causing all commands to be re-executed. It is not quite as easy to ignore directories in the FindFirstFile wrapper, so instead both fuse and Windows now report directory accesses, and they are dropped in add_node_to_list() if we know the tent is a directory.