Added t4076 - fix 'sed -i'.
When sed is run with the -i flag, it creates a temporary file using mode 000. If create() is not implemented, then fuse tries to do mknod(file, 000); open(file), which fails because it doesn't have permissions to open the file anymore. Tup needs to implement the create call so this occurs as a single operation and sed is happy.