Added t2128 - allow preloading other directories for readdir() in run-scripts.
Normally run-scripts can only readdir() in the current directory, since we have no idea what the state of other directories is (files may be built in them, or not). Now the preload keyword can be used to force parsing other Tupfiles that we want to look at, so tup can know the directory state when the script executes.