Fix t8063 - don't delete variant dirs when moved with the monitor.
It doesn't make a lot of sense to have different behavior in t8062/t8063 since one just involves the monitor. Also, it can be confusing for a user to try to make a backup of a variant directory by moving it away, and then seeing it get destroyed by the next 'tup upd'. Instead, just disable the variant and make all of the generated files now normal files.

This could also break the monitor, since the updater was removing directories that the monitor was watching. Strangely, I only see it in arch linux, but not ubuntu, fedora, centos, or gentoo.