Parser version 7: Disallow node-variables from referring to generated files.
Previously, a node-variable could refer to a generated file like so:

: |> ... |> a.txt
&var = a.txt

However, this becomes more difficult when referring to a generated file in another directory:

&var = ../foo/a.txt

Because we may not have parsed foo/Tupfile yet, so foo/a.txt doesn't exist. This causes the build to behave non-deterministically, which is not acceptable. We can change the node-variable lookup so that it uses input_pattern_to_nl() instead of get_tent_dt(), which would cause it to stop parsing the current Tupfile and parse foo/Tupfile first, similar to input files in :-rules. However, this then breaks t2104 due to a circular dependency among the Tupfiles. Since this test is most similar to how node-variables were intended to be used, it seems the best option is to not allow node-variables to point to generated files.