Added t4078, t4079, t4080, t4081 - files with exit failures.
Sometimes it is useful to move over the valid files that were created by a sub-process even if the sub-process fails. For example, in Latex when it outputs a log file and fails, we don't want to just leave it under .tup/tmp as a random numbered file, but rather put it into the tree in the expected location. These tests cover the various cases, such as when only some of the outputs exist, or we accidentally write a file we're not supposed to (which shouldn't be moved over).