Added t3058 - partial updates shouldn't update the whole group.
In t6055, logic was added to make sure that all output nodes of a command are kept as long as at least one is needed. However, if one of the outputs is a group, we don't want to include all commands pointing to that group. This test checks for the correct behavior when a partial update is used in a group.