Added t8076, t8077 - clean up empty Tupfiles correctly with variants.
Normally when a directory is removed in the src tree, it ends up being a ghost until the generated files and such are cleaned up, so the srcid is still valid for the corresponding variant directory. However, if there is just an empty Tupfile (or one with comments), then when the directory is removed we end up just deleting the node immediately. This means the srcid is now invalid, so later when we try to process the variant dir, we can't find its srcid. One option is to make the ghost check also check for srcid, but since that's not indexed we would have to rev the db. This implementation just sets the srcid to a special value (-2), so when we hit these in the parser we use that as a marker to clean up the variant directory as well.