Save both a full graph and trimmed graph when hitting a fatal circular dependency error.
The circular dependency may be discovered early in processing a large graph, so dumping the whole thing may not be useful in this case. We can trim the graph down to just the circular part by removing all nodes that either have no incoming links or no outgoing links, and then save the result. This should make the circular dependency more obvious to the user.