Parser version 8: forward sticky links to future nodes.
Instead of doing a graph search for transitive dependency detection, we can instead forward sticky links from command A to command B during the parse stage. This does more work up front and stores more data in the database, but it means we don't have to do unbounded searches later.