Added t6069 - check for incremental circular dependency with groups.
The problem here is that if the Tupfile is constructed incrementally, the build appears to succeed. However, doing a full build afterward causes a fatal tup error because of the circular dependency with the group. Fixing this was a bit tricky, since when we are outputting to a group, we don't know what groups are earlier than us in the chain without building a (possibly large) PDAG starting with the group.

The fix is to use the forwarded sticky links so that the command writing to the group already has all the input groups available to check.